Fix HTTP::REST Transport error if request body is empty
Previously if you'd perform an request using the HTTP::REST backend that
had a empty request body (i. e. { Data => {}, }), it would fail with
this error message:

> Not a SCALAR reference at /opt/otrs/bin/cgi-bin/../../Kernel/cpan-lib/LWP/Protocol/http.pm line 256.

This is caused by REST::Client missing its body argument. The solution
is to pass undef as a body argument if our body turns out to be empty.